Fixed bug where files weren't completely written out

+++ include/io.h
@@ -4,4 +4,4 @@
 #include <stdlib.h>
 void printFileError(const char *filename, const char *invocation);
 size_t readFile(const char *filename, u8 **buffer);
-bool writeFile(const char *filename, const u8 *buffer, u8 bufferLength);
+bool writeFile(const char *filename, const u8 *buffer, size_t bufferLength);
